k8s: add k8s support for runu using raw-socket netfd
k8s requires to detect the IP address assinged to a container, thus
lkl.json is needed to be auto-generated during the boot process.

To do that, runu now uses the IP address assinged by containerd (?) by
deleting the address in the namsepace (from vethXXX).  But we had handful of
tricks to achive:

- has to disable tx csum offload at host veth (need to implement vnethdr
for raw-socket backend)
- to disable tx csum in host veth, we used the interface with the largest
ifindex to disable tx csum (by ethtool ioctl)
- delete IP address of vethXXX in namespace to avoid address collision
(the way gvisor does)
- now ifname in lkl.json indicate the type of net backend (tapXXX or
ethXXX to distinguish)
- gateway address is .0.1 of the interface address (need to obtain from
getIPNetAddr)
- pause container (in k8s case) is treated to not touch the IP address
- no IPv6 address support

Those tricks should be cleaned up in the future commit.

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>