shim: add reaper implementation for darwin
This implementation was originally in the containerd port for darwin, but
now move to shim implementation to reduce the big changes to the upstream
containerd.

darwin has the different behavior of handling orphaned child
processes: when parent process dies, children process cannot be a
child of grandmother's process (no prctl(PR_SET_CHILD_SUBREAPER)
support as Linux/FreeBSD does).  Thus, wait4(2) can't catch a termination
of children.  We put a workaround for this.

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>